OP-7470 - added new validator for single frame
