+++ openpype/hosts/fusion/plugins/publish/validate_instance_frame_range.py
@@ -20,21 +20,6 @@ class ValidateInstanceFrameRange(pyblish.api.InstancePlugin):
         render_start = instance.data["frameStartHandle"]
         render_end = instance.data["frameEndHandle"]
 
-        if instance.data["family"] == "image":
-            too_many_frames = (isinstance(instance.data["expectedFiles"], list)
-                               and len(instance.data["expectedFiles"]) > 1)
-
-            if render_end - render_start > 0 or too_many_frames:
-                desc = "Trying to render {}-{} and expected files: {}.".format(
-                    render_start, render_end, instance.data["expectedFiles"])
-                desc += ("<br><br>Either change frame range based on "
-                         "`Frame range source` (eg. asset, node or timeline)"
-                         " or use `Render (saver)` creator")
-                raise PublishValidationError(
-                    title="Frame range outside of comp range",
-                    message=desc,
-                    description=desc
-                )
         if render_start < global_start or render_end > global_end:
 
             message = (
